projects: iio_demo: Fix 64-bit platforms incompatibility
The address was limited to 32 bits, independent of the platform.

+++ projects/iio_demo/src/app/main.c
@@ -82,8 +82,8 @@ static uint16_t loopback_buffs[DEMO_CHANNELS][SAMPLES_PER_CHANNEL];
 static uint8_t in_buff[MAX_SIZE_BASE_ADDR];
 static uint8_t out_buff[MAX_SIZE_BASE_ADDR];
 
-#define DAC_DDR_BASEADDR	((uint32_t)out_buff)
-#define ADC_DDR_BASEADDR	((uint32_t)in_buff)
+#define DAC_DDR_BASEADDR	out_buff
+#define ADC_DDR_BASEADDR	in_buff
 
 #endif
